refactor(scroller): creates some helpful utils
